+++ src/App.js
@@ -1,6 +1,6 @@
-import Expenses from "./components/Expenses";
+import Expenses from "./components/Expenses/Expenses";
 
-function App() {
+const App = () => {
   const expenses = [
     {
       id: "e1",
@@ -34,6 +34,6 @@ function App() {
       <Expenses items={expenses}/>
     </div>
   );
-}
+};
 
 export default App;
