+++ src/components/Expenses/ExpenseDate.js
@@ -1,6 +1,6 @@
 import './ExpenseDate.css'
 
-function ExpenseDate(props) {
+const ExpenseDate = props => {
   const month = props.date.toLocaleString('en-US', { month: 'long'});
   const year = props.date.getFullYear();
   const day = props.date.toLocaleString('en-US', { day: '2-digit'})
