+++ src/components/Expenses/ExpenseItem.css
@@ -2,10 +2,8 @@
   display: flex;
   justify-content: space-between;
   align-items: center;
-  box-shadow: 0 2px 8px rgba(0, 0, 0, 0.25);
   padding: 0.5rem;
   margin: 1rem 0;
-  border-radius: 12px;
   background-color: #4b4b4b;
 }
